automation_tools/requirement.txt
art==6.1
bashplotlib==0.6.5
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.5
cx-Oracle==8.3.0
numpy==1.26.1
pandas==2.1.1
paramiko==3.3.1
pip==23.1.2
plotext==5.2.8
pycparser==2.21
pycryptodome==3.19.0
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
setuptools==65.5.0
six==1.16.0
termplot==0.0.2
termplotlib==0.3.9
tzdata==2023.3